requirements.txt
alembic==1.9.1
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
configparser==5.3.0
cryptography==38.0.4
decorator==5.1.1
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
exceptiongroup==1.1.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
outcome==1.2.0
packaging==23.2
pi==0.1.2
Pillow==9.4.0
pycparser==2.21
pydub==0.25.1
PyMySQL==1.0.2
PySocks==1.7.1
python-barcode==0.14.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0.1
ratelim==0.1.6
requests==2.28.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.14
visitor==0.1.3
waitress==2.1.2
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.11.0
